feat: dependencies under views are moved to seperate deps file
